+++ cli/cli/config/config.go
@@ -1,0 +1,11 @@
+package config
+
+type Config interface {
+	HasMetricsConsentPromptBeenDisplayed() bool
+	MetricsConsentPromptHasBeenDisplayed()
+	HasUserAcceptedSendingMetrics() bool
+	UserAcceptSendingMetrics()
+	UserDoNotAcceptSendingMetrics()
+	Save() error
+	String() string
+}